Add a new averaging_engine implemented using cpp.
Configure the use of this engine with the phil parameter
averaging_engine=python, cpp, or both.  If both is selected,
the original python code will be ran, then the cpp version will be ran, and
finally the results will be checked to assure binary identity.

The original behavior (python) is currently the default.

Logs created using flag_output_verbose for both engines should also be identical,
use diff to verify.